Crypto-main/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
tortoise-orm==0.19.2
passlib==1.7.4
python-dotenv==0.21.1
PyJWT==2.6.0
pydantic==1.10.4
python-multipart==0.0.5
bcrypt==4.0.1
aiosqlite==0.17.0
